requirements.txt
alembic==1.14.0
blinker==1.9.0
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.7
MarkupSafe==3.0.2
msgspec==0.18.6
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
